Move /site-updates/* controllers into own module
