+++ weasyl/controllers/admin.py
@@ -3,7 +3,7 @@ from pyramid.response import Response
 
 from libweasyl import staff
 
-from weasyl import moderation, profile, siteupdate
+from weasyl import moderation, profile
 from weasyl.error import WeasylError
 from weasyl.controllers.decorators import admin_only
 from weasyl.controllers.decorators import token_checked
@@ -16,73 +16,6 @@ import weasyl.define as d
 @admin_only
 def admincontrol_(request):
     return Response(d.webpage(request.userid, "admincontrol/admincontrol.html", title="Admin Control Panel"))
-
-
-_BLANK_SITE_UPDATE = {
-    'updateid': None,
-    'title': "",
-    'content': "",
-    'wesley': False,
-}
-
-
-@admin_only
-def admincontrol_siteupdate_get_(request):
-    return Response(d.webpage(request.userid, "siteupdates/form.html", (_BLANK_SITE_UPDATE,), title="Submit Site Update"))
-
-
-@admin_only
-@token_checked
-def admincontrol_siteupdate_post_(request):
-    title = request.POST["title"].strip()
-    content = request.POST["content"].strip()
-    wesley = "wesley" in request.POST
-
-    if not title:
-        raise WeasylError("titleInvalid")
-
-    if not content:
-        raise WeasylError("contentInvalid")
-
-    updateid = siteupdate.create(
-        userid=request.userid,
-        title=title,
-        content=content,
-        wesley=wesley,
-    )
-
-    raise HTTPSeeOther(location="/site-updates/%d" % (updateid,))
-
-
-@admin_only
-def site_update_edit_(request):
-    updateid = int(request.matchdict['update_id'])
-    update = siteupdate.select_view(updateid)
-    return Response(d.webpage(request.userid, "siteupdates/form.html", (update,), title="Edit Site Update"))
-
-
-@admin_only
-@token_checked
-def site_update_put_(request):
-    updateid = int(request.matchdict['update_id'])
-    title = request.POST["title"].strip()
-    content = request.POST["content"].strip()
-    wesley = "wesley" in request.POST
-
-    if not title:
-        raise WeasylError("titleInvalid")
-
-    if not content:
-        raise WeasylError("contentInvalid")
-
-    siteupdate.edit(
-        updateid=updateid,
-        title=title,
-        content=content,
-        wesley=wesley,
-    )
-
-    return HTTPSeeOther(location="/site-updates/%d" % (updateid,))
 
 
 @admin_only
